requirements.txt
pandas==1.5.3
python-dotenv==1.0.0
requests==2.29.0
tqdm==4.64.1
